[4633] Move shared course attributes into concern for hesa and hpitt
